Implement drawing and starting positions
