+++ TeamCode/src/main/java/org/firstinspires/ftc/teamcode/pedroPathing/Tuning.java
@@ -10,9 +10,6 @@ import static org.firstinspires.ftc.teamcode.pedroPathing.Tuning.telemetryM;
 import com.bylazar.configurables.PanelsConfigurables;
 import com.bylazar.configurables.annotations.Configurable;
 import com.bylazar.configurables.annotations.IgnoreConfigurable;
-import com.bylazar.field.FieldManager;
-import com.bylazar.field.PanelsField;
-import com.bylazar.field.Style;
 import com.bylazar.telemetry.PanelsTelemetry;
 import com.bylazar.telemetry.TelemetryManager;
 import com.pedropathing.follower.Follower;
@@ -1221,151 +1218,5 @@ class Circle extends OpMode {
 		if (follower.atParametricEnd()) {
 			follower.followPath(circle);
 		}
-	}
-}
-
-/**
- * This is the Drawing class. It handles the drawing of stuff on Panels Dashboard, like the robot.
- *
- * @author Lazar - 19234
- * @version 1.1, 5/19/2025
- */
-class Drawing {
-	public static final double ROBOT_RADIUS = 9; // woah
-	private static final FieldManager panelsField = PanelsField.INSTANCE.getField();
-	
-	private static final Style robotLook = new Style(
-			"", "#3F51B5", 0.75
-	);
-	private static final Style historyLook = new Style(
-			"", "#4CAF50", 0.75
-	);
-	
-	/**
-	 * This prepares Panels Field for using Pedro Offsets
-	 */
-	public static void init() {
-		panelsField.setOffsets(PanelsField.INSTANCE.getPresets().getPEDRO_PATHING());
-	}
-	
-	/**
-	 * This draws everything that will be used in the Follower's telemetryDebug() method. This takes
-	 * a Follower as an input, so an instance of the DashbaordDrawingHandler class is not needed.
-	 *
-	 * @param follower Pedro Follower instance.
-	 */
-	public static void drawDebug(Follower follower) {
-		if (follower.getCurrentPath() != null) {
-			drawPath(follower.getCurrentPath(), robotLook);
-			Pose closestPoint = follower.getPointFromPath(follower.getCurrentPath().getClosestPointTValue());
-			drawRobot(new Pose(closestPoint.getX(), closestPoint.getY(), follower.getCurrentPath().getHeadingGoal(follower.getCurrentPath().getClosestPointTValue())), robotLook);
-		}
-		drawPoseHistory(follower.getPoseHistory(), historyLook);
-		drawRobot(follower.getPose(), historyLook);
-		
-		sendPacket();
-	}
-	
-	/**
-	 * This draws a robot at a specified Pose with a specified
-	 * look. The heading is represented as a line.
-	 *
-	 * @param pose  the Pose to draw the robot at
-	 * @param style the parameters used to draw the robot with
-	 */
-	public static void drawRobot(Pose pose, Style style) {
-		if (pose == null || Double.isNaN(pose.getX()) || Double.isNaN(pose.getY()) || Double.isNaN(pose.getHeading())) {
-			return;
-		}
-		
-		panelsField.setStyle(style);
-		panelsField.moveCursor(pose.getX(), pose.getY());
-		panelsField.circle(ROBOT_RADIUS);
-		
-		Vector v = pose.getHeadingAsUnitVector();
-		v.setMagnitude(v.getMagnitude() * ROBOT_RADIUS);
-		double x1 = pose.getX() + v.getXComponent() / 2, y1 = pose.getY() + v.getYComponent() / 2;
-		double x2 = pose.getX() + v.getXComponent(), y2 = pose.getY() + v.getYComponent();
-		
-		panelsField.setStyle(style);
-		panelsField.moveCursor(x1, y1);
-		panelsField.line(x2, y2);
-	}
-	
-	/**
-	 * This draws a robot at a specified Pose. The heading is represented as a line.
-	 *
-	 * @param pose the Pose to draw the robot at
-	 */
-	public static void drawRobot(Pose pose) {
-		drawRobot(pose, robotLook);
-	}
-	
-	/**
-	 * This draws a Path with a specified look.
-	 *
-	 * @param path  the Path to draw
-	 * @param style the parameters used to draw the Path with
-	 */
-	public static void drawPath(Path path, Style style) {
-		double[][] points = path.getPanelsDrawingPoints();
-		
-		for (int i = 0; i < points[0].length; i++) {
-			for (int j = 0; j < points.length; j++) {
-				if (Double.isNaN(points[j][i])) {
-					points[j][i] = 0;
-				}
-			}
-		}
-		
-		panelsField.setStyle(style);
-		panelsField.moveCursor(points[0][0], points[0][1]);
-		panelsField.line(points[1][0], points[1][1]);
-	}
-	
-	/**
-	 * This draws all the Paths in a PathChain with a
-	 * specified look.
-	 *
-	 * @param pathChain the PathChain to draw
-	 * @param style     the parameters used to draw the PathChain with
-	 */
-	public static void drawPath(PathChain pathChain, Style style) {
-		for (int i = 0; i < pathChain.size(); i++) {
-			drawPath(pathChain.getPath(i), style);
-		}
-	}
-	
-	/**
-	 * This draws the pose history of the robot.
-	 *
-	 * @param poseTracker the PoseHistory to get the pose history from
-	 * @param style       the parameters used to draw the pose history with
-	 */
-	public static void drawPoseHistory(PoseHistory poseTracker, Style style) {
-		panelsField.setStyle(style);
-		
-		int size = poseTracker.getXPositionsArray().length;
-		for (int i = 0; i < size - 1; i++) {
-			
-			panelsField.moveCursor(poseTracker.getXPositionsArray()[i], poseTracker.getYPositionsArray()[i]);
-			panelsField.line(poseTracker.getXPositionsArray()[i + 1], poseTracker.getYPositionsArray()[i + 1]);
-		}
-	}
-	
-	/**
-	 * This draws the pose history of the robot.
-	 *
-	 * @param poseTracker the PoseHistory to get the pose history from
-	 */
-	public static void drawPoseHistory(PoseHistory poseTracker) {
-		drawPoseHistory(poseTracker, historyLook);
-	}
-	
-	/**
-	 * This tries to send the current packet to FTControl Panels.
-	 */
-	public static void sendPacket() {
-		panelsField.update();
 	}
 }
